Subject: templates: add multinode_measurements template
In order to submit automatically a job using Multinode Lava protocol to
measure the power consumption of a board while runing tests, we
implement a template that uses LXC as a host and a board as a target.
In this template the LXC host will use the pyacmecapture.py script to measure
power consumption of the board during the test section.

The host and the target start the jobs in parallel then the host will
wait the guest to start the test section and if both of them are
synchronised, the guest will execute some shell commands and in the same
time the host will measure the power consumption of the board.

To execute the pyacmecapture.py script we need the IP address of the
ACME and the number of the Slot , those information are added in the
device dict as static info then we extract them in the job using lavacli tool
Once the measurements are finished, the output files will be uploaded
in file server that we have to indicate its path in as environnement
variable in lava-master/etc/lava-server/env.yaml

Finally we obtain 2 output files one is .csv file which is RAW DATA and
the .txt file is an ACME SUMMARY.

In this template we include target_test.jinja2 which contains the test
section of the target.
And the LXC test section is written in a script on my Github then
 it will be integrated on Baylibre Github.

Signed-off-by: Emna Trabelsi <etrabelsi@baylibre.com>